+++ packages/stats/gbstats/models/statistics.py
@@ -1,10 +1,8 @@
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from typing import List, Optional, Union
+from typing import Union
 
 import numpy as np
-
-from gbstats.shared.constants import DifferenceType
+from pydantic.dataclasses import dataclass
 
 
 @dataclass
@@ -202,39 +200,3 @@ TestStatistic = Union[
     RegressionAdjustedStatistic,
     RatioStatistic,
 ]
-
-
-# Data class for test config
-@dataclass
-class BaseConfig:
-    difference_type: DifferenceType = DifferenceType.RELATIVE
-    traffic_proportion_b: float = 1
-    phase_length_days: float = 1
-
-
-# Data classes for the results of tests
-@dataclass
-class Uplift:
-    dist: str
-    mean: float
-    stddev: float
-
-
-@dataclass
-class TestResult:
-    expected: float
-    ci: List[float]
-    uplift: Uplift
-
-
-@dataclass
-class BayesianTestResult(TestResult):
-    chance_to_win: float
-    risk: List[float]
-    error_message: Optional[str] = None
-
-
-@dataclass
-class FrequentistTestResult(TestResult):
-    p_value: float
-    error_message: Optional[str] = None
