+++ packages/stats/pyproject.toml
@@ -29,6 +29,7 @@ pandas = "^1.5.1"
 scipy = "^1.9.2"
 numpy = "^1.23.4"
 nbformat = "^5.8.0"
+pydantic = "^2.5.3"
 
 [tool.poetry.group.dev.dependencies]
 certifi = "2022.12.7"
