+++ packages/stats/tests/bayesian/test_tests.py
@@ -5,6 +5,7 @@ from unittest import TestCase, main as unittest_main
 import numpy as np
 
 from gbstats.bayesian.tests import (
+    BayesianTestResult,
     BetaPrior,
     BinomialBayesianABTest,
     BinomialBayesianConfig,
@@ -12,12 +13,11 @@ from gbstats.bayesian.tests import (
     GaussianPrior,
     GaussianBayesianConfig,
 )
-from gbstats.shared.models import (
-    BayesianTestResult,
-    Uplift,
+from gbstats.models.statistics import (
     ProportionStatistic,
     SampleMeanStatistic,
 )
+from gbstats.models.tests import Uplift
 
 DECIMALS = 5
 round_ = partial(np.round, decimals=DECIMALS)
